social_sentiment_pipeline/requirements.txt
# Core dependencies
apache-airflow==2.7.0
pyspark==3.4.1
kafka-python==2.0.2
clickhouse-driver==0.2.6

# ML and NLP
transformers==4.33.2
torch==2.0.1
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3

# Twitter API
tweepy==4.14.0

# Data processing
pyarrow==12.0.1
fastparquet==0.8.3

# Monitoring and logging
prometheus-client==0.17.1
structlog==23.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0

# Development
pytest==7.4.0
black==23.7.0
flake8==6.0.0